+++ changelog.md
@@ -5,6 +5,7 @@
 ### fix
 
 - discord rich presence not inmediately updated after settings change.
+- inconsistency on styles of dock and quick settings widgets.
 
 ## [2.6.10]